added some collision and physics
I added physics and collision to the test block that i placed in the world, all I have to figure out now is why there is some sort of gust of wind that is blowing me away. Also the MinecraftCamera class is starting to look like it might just need to be contained in the physics class, it seems kind of useless. We'll see