Fixed the jumping and shift button
Game now can only jump once and if you press shift the speed of the character slows down. I was able to get the collision detector to work by adding an invinsible spatial to the character that the collision listener could read and then every time the collision was detected it would set a boolean to true and when the space bar was pressed it would see if that was true to jump and once the space bar was let go the boolean was set to false so the person couldnt hit the space bar again while in the air.