made physics more efficient, made blocks not fall through eachother, and added a crosshair
I created a collision group and made it only contain the player at first to reduce the amount of collision events calculated then I realized that i can fix blocks falling through by calculating collisions so I made it that anything with weight (which won't be a lot) will be added to the collision group that gets calculated. Added to main a getter for the AppSettings class. Also added a basic crosshair, will probably do more with it later