wrote two classes, one to handle creating blocks, the other to hold what kind of blocks there are
BlockDatabase takes in an id and coordinate to search what block to create and where to place it. It passes that info to the block class who handles adding collision to it and rendering it. In main I wrote a temporary piece of code to see it in work. tomorrow I will work on destroying blocks