Spent this time moving code around
A lot was being done in the physics class that wasn't even physics related so I moved them to their respective classes, I moved the key pressing to keyMapping. I moved player spawn and destroying and creating blocks to world gen. I also moved all the calculations for the player to the camera class so now it's actually useful