Tried to improve performance but failed and tried to add chunk deleting but it's crashing the jvm
I moved the mesh creation when there is a chunk update to a different thread thinking that was the source of lag when breaking a bunch of blocks at once but that didn't fix anything. I then tried to implement chunk deleting but for some reason that just crashes the jvm which is really annoying.