Fixed the saving
Kind of a weird issue but basically the way the saving works is when a block is broken or placed the method that handles that gets called and checks the collision results and either breaks or places the blocks at a certain coordinate, in that same method it calls the save methods and somehow while the method that handles the breaking and placing is still alive the collision results change so it was writing the wrong coordinates to the file. one variable later and all better.